rename sdk files and add process quiting
